+++ src/main/java/me/zeroeightsix/kami/gui/rgui/component/Button.kt
@@ -1,17 +1,17 @@
 package me.zeroeightsix.kami.gui.rgui.component
 
-import me.zeroeightsix.kami.util.math.Vec2d
+import me.zeroeightsix.kami.util.math.Vec2f
 
-open class Button(override var name: String, val action: () -> Unit) : Slider(name, 0.0) {
+open class Button(override var name: String, val action: () -> Unit) : Slider(name, 0.0f) {
 
-    override fun onClick(mousePos: Vec2d, buttonId: Int) {
+    override fun onClick(mousePos: Vec2f, buttonId: Int) {
         super.onClick(mousePos, buttonId)
-        value = 1.0
+        value = 1.0f
         action()
     }
 
-    override fun onRelease(mousePos: Vec2d, buttonId: Int) {
+    override fun onRelease(mousePos: Vec2f, buttonId: Int) {
         super.onRelease(mousePos, buttonId)
-        value = 0.0
+        value = 0.0f
     }
 }
